fix build error

+++ net/http/include/Http2PacketWriterImpl.hpp
@@ -13,7 +13,7 @@ DECLARE_CLASS(Http2PacketWriterImpl) IMPLEMENTS(HttpPacketWriter) {
 
 public:
     explicit _Http2PacketWriterImpl(Http2StreamController);
-    int write(HttpPacket);
+    long write(HttpPacket);
     ByteArray data(HttpPacket);
 
 private:
